Reorganize skills into folder
