+++ codex-rs/core/src/skills/mod.rs
@@ -1,0 +1,7 @@
+pub mod model;
+pub mod render;
+pub mod loader;
+
+pub use loader::{load_skills, SkillError, SkillLoadOutcome};
+pub use model::{SkillMetadata};
+pub use render::render_skills_section;
